refactor(llm): split LLM.cpp -- init + layer-topology into LLM_init.cpp (slice 2)
De-inlines the initialization + topology-construction methods out of LLMImpl.hpp
into LLM_init.cpp: Init, Deinit, init_groups_from_model, init_layer_groups,
init_layer_types, init_shared_kv_source_layers, distributeModels (~912 lines).
LLMImpl.hpp 4757->3853; LLM_init.cpp 922.

distributeModels is #ifdef USE_AXCL-guarded (multi-card, AXCL-only); the def in
LLM_init.cpp carries the same guard so the non-AXCL build has no orphan definition.

Pure code-move, behavior-preserving. Verified on AX650N .51 BOTH backends:
- build_650 (non-AXCL) + build_x86 (AXCL) both compile+link clean
- qwen3-0.6b: 5-turn output + KV-content fingerprint BYTE-IDENTICAL to pre-split
  baseline (same hashes as slice 1); Qwen3-VL-2B image path correct + deterministic

Next slice: decode (the two Run()s, hot path) -> LLM_decode.cpp.
